backend/requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14
django-cors-headers>=4.3
gunicorn>=21.2
dj-database-url>=2.1
psycopg2-binary>=2.9
